changed the saved file
svn path=/trunk/iCub/contrib/src/eMorph/; revision=15007

+++ src/aexGrabber/include/iCub/device2yarp.h
@@ -59,7 +59,7 @@ struct aer {
 };
 
 struct atom {
-    u32 data;
+    u16 data;
 };
